Use new json serializer in rabbitmq contetxs
