+++ back/Dockerfile.dev
@@ -20,4 +20,4 @@ EXPOSE 5000
 ENV DOTNET_USE_POLLING_FILE_WATCHER 1
 # HEALTHCHECK --interval=5s CMD curl --fail http://localhost:5000/health || exit
 HEALTHCHECK CMD true
-ENTRYPOINT ["dotnet", "watch", "run", "--no-restore", "--project", "/app/src/Kyoo.Host"]
+ENTRYPOINT ["dotnet", "watch", "--non-interactive", "run", "--no-restore", "--project", "/app/src/Kyoo.Host"]
